cf_extract/requirements.txt
functions-framework
google-cloud-storage
google-cloud-bigquery
requests
python-dotenv